requirements.txt
streamlit
openai
requests
beautifulsoup4
python-dotenv
tldextract
